+++ envsec/auth/README.md
@@ -1,0 +1,3 @@
+# Auth
+2023-08-21: This package is going to be refactored into it's own go module. Don't introduce
+extra dependencies here.
