--- pkg/go.mod
@@ -1,20 +1,0 @@
-module go.jetpack.io/pkg
-
-go 1.20
-
-require (
-	github.com/MicahParks/keyfunc/v2 v2.1.0
-	github.com/golang-jwt/jwt/v5 v5.0.0
-	github.com/pelletier/go-toml/v2 v2.0.7
-	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8
-	github.com/pkg/errors v0.9.1
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/kr/pretty v0.3.0 // indirect
-	github.com/rogpeppe/go-internal v1.10.0 // indirect
-	github.com/stretchr/testify v1.8.4 // indirect
-	golang.org/x/sys v0.8.0 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-)
